requirements.txt
pytest
retry
pytest-html
flake8
pyyaml
selenium
webdriver-manager